+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Added
+- WSS2S class with support for callRequestTokenS2S and callRequestTokenS2S methods
+- WSS2S class with support for callCheckCartaS2S, callVerifyCardS2S, callSettleS2S, callRefundS2S and callDeleteS2S methods
 
 ## [1.3.0] - 2015-11-06
 ### Changed